refactor: planner api  단일, 연속 수정/삭제 업데이트 1
refactor: planner api  단일, 연속 수정/삭제 업데이트 1

+++ src/planners/planners.controller.ts
@@ -47,20 +47,37 @@ export class PlannersController {
   async update(
     @Req() req: any,
     @Param('plannerId') plannerId: string,
-    @Body() updatePlannerDto: PlannerDto
+    @Body() plannerDto: PlannerDto
   ): Promise<any> {
     const userId = req.user._id;
-    return this.plannersService.updatePlan(userId, plannerId, updatePlannerDto);
+    const { isContinuous, ...updatePlannerDto } = plannerDto;
+
+    if (isContinuous) {
+      // 단일 수정
+    } else {
+      // 연속 수정
+      return this.plannersService.updatePlanCascade(
+        userId,
+        plannerId,
+        updatePlannerDto
+      );
+    }
   }
 
   @UseGuards(AuthGuard('jwt'))
   @Delete(':plannerId')
   async delete(
     @Req() req: any,
-    @Param('plannerId') plannerId: string
+    @Param('plannerId') plannerId: string,
+    @Query('isContinuous') isContinuous: boolean
   ): Promise<Planner> {
     const userId = req.user._id;
-    return this.plannersService.deletePlan(userId, plannerId);
+    if (isContinuous) {
+      // 단일 수정
+    } else {
+      // 연속 삭제
+      return this.plannersService.deletePlan(userId, plannerId);
+    }
   }
 
   @UseGuards(AuthGuard('jwt'))
